+++ app/scripts/LeftTrackModifier.js
@@ -154,12 +154,11 @@ class LeftTrackModifier {
     this.xScale(newXScale);
     this.yScale(newYScale);
 
-    if (this.originalTrack.refreshTiles) {
-      // some tracks don't have refreshTiles (e.g. PixiTrack)
-      this.originalTrack.refreshTiles();
-    }
-
     if (this.originalTrack.leftTrackZoomed) {
+      if (this.originalTrack.refreshTiles) {
+        // some tracks don't have refreshTiles (e.g. PixiTrack)
+        this.originalTrack.refreshTiles();
+      }
       // the track implements its own left-oriented zooming and scrolling
       this.originalTrack.leftTrackZoomed(newXScale, newYScale, k, tx, ty);
       this.originalTrack.draw();
@@ -181,7 +180,17 @@ class LeftTrackModifier {
       this.originalTrack.pMobile.position.y = this.originalTrack.position[1];
     }
 
-    this.originalTrack.draw();
+    if (this.originalTrack.leftTrackDraw) {
+      // if the track implements leftTrackDraw we just redraw the track and
+      // won't call the track's zoomed method
+      if (this.originalTrack.refreshTiles) {
+        this.originalTrack.refreshTiles();
+      }
+      this.originalTrack.leftTrackDraw();
+      return;
+    }
+
+    this.originalTrack.zoomed(this.xScale(), this.yScale());
   }
 
   zoomedY(yPos, kMultiplier) {